% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.gather.R
\name{assignment.collect}
\alias{assignment.collect}
\title{Collect homeworks submitted to GitHub classroom}
\usage{
assignment.collect(organization, GHpattern, path)
}
\arguments{
\item{organization}{Name of GitHub organization}

\item{GHpattern}{Pattern in GitHub repos for the target assignment}

\item{path}{Where should the repos be cloned locally?}
}
\description{
This function takes the name of an organization, a pattern in student's repos
and download all the matching repos in the organization to a particular folder.
}
