% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.expand.R
\name{assignment.expand}
\alias{assignment.expand}
\title{Generate student's and full version of the homework}
\usage{
assignment.expand(path = ".")
}
\description{
This function takes a pre-defined HW template and generates a full
homework with an autograding system
}
