% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.grades.R
\name{assignment.grades}
\alias{assignment.grades}
\title{Generate student's and full version of the homework}
\usage{
assignment.grades(createPDFs = TRUE)
}
\arguments{
\item{createPDFs}{Should we use \code{wkhtmltopdf} to create PDF
files from student's submissions?}
}
\description{
This function extracts grades from submitted assignments. It also
moves all the HTML files from each student repo to a main folder where
the instructor can check the rendered submissions.If you're interested in
using \code{wkhtmltopdf} to transform HTML files into PDFs, and submit comments
in PDF files, that's also an option. However, please make sure that \code{wkhtmltopdf}
is installed to the path. Instructors can also just give feedback directly in
github or using any other approach they decide is suitable given their expectations.
}
