% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.template.R
\name{assignment.template}
\alias{assignment.template}
\title{Create a homework template for gradeR}
\usage{
assignment.template(path = ".")
}
\arguments{
\item{path}{Where should the sample directory be exported to?}
}
\description{
This function generates the structure of a basic autograding system.
}
