% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor.R
\name{autograder_addin_configure}
\alias{autograder_addin_configure}
\title{RStudio Addin: Configure Autograder}
\usage{
autograder_addin_configure()
}
\description{
Opens a dialog for configuring autograder (used by RStudio Addins menu).
}
\keyword{internal}
