% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor.R
\name{autograder_check_build_tools}
\alias{autograder_check_build_tools}
\title{Check if R build tools are available}
\usage{
autograder_check_build_tools()
}
\value{
Logical TRUE if build tools are available
}
\description{
Verifies that the system has the required tools to build R packages
with C++ code (Rtools on Windows, Xcode on Mac, gcc on Linux).
}
\examples{
\dontrun{
autograder_check_build_tools()
}

}
\keyword{internal}
