% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{autograder_diagnose}
\alias{autograder_diagnose}
\title{Diagnose Common Issues}
\usage{
autograder_diagnose()
}
\value{
A list with diagnostic results.
}
\description{
Runs a comprehensive diagnostic check to identify common issues with
the autograder setup.
}
\examples{
\dontrun{
autograder_diagnose()
}

}
\keyword{internal}
