% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced.R
\name{autograder_log_clear}
\alias{autograder_log_clear}
\title{Clear log history}
\usage{
autograder_log_clear()
}
\value{
Invisibly returns TRUE.
}
\description{
Clears the in-memory log history.
}
\keyword{internal}
