% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_for_special_functions}
\alias{check_for_special_functions}
\title{check_for_special_functions}
\usage{
check_for_special_functions(filename, func_vector)
}
\arguments{
\item{filename}{a character string which is a file name}

\item{func_vector}{a character vector of special functions. It is best if this is just one or two functions, especially \code{for} and \code{while}}
}
\value{
a numeric count of violations for the specified functions
}
\description{
A generic function to check whether specific special functions have been used
special functions include \code{if}, \code{for}, \code{while}, \code{repeat}, \code{break}, \code{next}, \code{return}, \code{function}, \code{quote}, \code{switch}, \verb{\{}, \code{(}
This link has the updated list of special function names
https://cran.r-project.org/doc/manuals/r-release/R-ints.html#g_t_002eInternal-vs-_002ePrimitive
}
\examples{
try(check_for_special_functions())
}
