% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{dollar_sign_counter}
\alias{dollar_sign_counter}
\title{Count the number of dollar signs used.}
\usage{
dollar_sign_counter(filename)
}
\arguments{
\item{filename}{a character string which is a file name}
}
\value{
a count of dollar sign, \code{$}, uses, e.g., to extract elements of a \code{list} (and thus \code{data.frame})
}
\description{
This function takes an R scripts and
outputs a count of dollar sign uses
works within function calls, too
}
\examples{
example_script("generic_script") |>
  dollar_sign_counter()
}
