% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_autograder.R
\name{.apply_autograder}
\alias{.apply_autograder}
\title{Assing value to score}
\usage{
.apply_autograder()
}
\value{
internal for each check function
}
\description{
Assing value to score
}
\details{
\if{html}{\out{<div class="sourceCode {asciicast cli-appauto}">}}\preformatted{.scores <- rep(-1, times = 1)
.check_q1 <-
  function() \{
    .problem_number <<- 1
    .autograder <<-
      function()\{
        if (is.character(q1)) return(c(value = -1, message = "Enter a raw number."))
        if (q1 == ncol(iris)) return(c(value = 0, message = "Calculating number of cols"))
        if (q1 == nrow(iris)) return(c(value = 1, message = paste("Correct!", praise::praise()) ))
     return(c(value = 0, message = "Wrong. Please try again."))
   \}
 .apply_autograder()
\}
q1 <- "YOUR ANSWER HERE"
.check_q1()
q1 <- ncol(iris)
.check_q1()
q1 <- nrow(iris)
.check_q1()
.score_print()
}\if{html}{\out{</div>}}
}
