% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_objects.R
\name{.check_items}
\alias{.check_items}
\title{Compare homework answers to an answer key}
\usage{
.check_items(key_items, key_objects, hw_objects)
}
\arguments{
\item{hw}{named list containing the objects created in the homework script.
Note that any named elements that don't have a corresponding named element
in `key` are ignored.}

\item{key}{named list containing the objects that represent the correct
answers. All elements in this list are compared to the similarly named
elements in `hw`.}
}
\value{
a `data.frame` with one row per name in `key`. The columns are:
  * `name`: name of the list element
  * `obj_status`: either "correct" (homework answers is the same as the key),
    "incorrect" (homework answer is different from the key), or "missing"
    (object is not present in `hw`)
  * `error`: If the answer was incorrect, contains the output from the
    comparison
}
\description{
Given a list representing homework answers and a list containing
  the correct answers, compares the correct answers to the homework answers
}
