% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_audit_logging_enabled}
\alias{.cpp_audit_logging_enabled}
\title{Check if audit logging is enabled (C++)}
\usage{
.cpp_audit_logging_enabled()
}
\value{
TRUE if enabled
}
\description{
Check if audit logging is enabled (C++)
}
\keyword{internal}
