% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_compare_set}
\alias{.cpp_compare_set}
\title{Unordered set comparison with sorting (C++)}
\usage{
.cpp_compare_set(actual, expected, tolerance = 1e-10)
}
\arguments{
\item{actual}{Vector from student}

\item{expected}{Expected vector}

\item{tolerance}{Numeric tolerance for numeric vectors}
}
\value{
TRUE if sets contain same elements
}
\description{
Unordered set comparison with sorting (C++)
}
\keyword{internal}
