% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_decrypt_url}
\alias{.cpp_decrypt_url}
\title{Decrypt URL}
\usage{
.cpp_decrypt_url(encrypted_url)
}
\arguments{
\item{encrypted_url}{Encrypted URL string}
}
\value{
Decrypted URL
}
\description{
Decrypts an encrypted URL.
}
\keyword{internal}
