% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_encrypt_url}
\alias{.cpp_encrypt_url}
\title{Encrypt URL}
\usage{
.cpp_encrypt_url(url)
}
\arguments{
\item{url}{URL to encrypt}
}
\value{
Encrypted URL string
}
\description{
Encrypts a URL for secure storage.
}
\keyword{internal}
