% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_fetch_function_content}
\alias{.cpp_fetch_function_content}
\title{Securely fetch function content (URL hidden)}
\usage{
.cpp_fetch_function_content(function_name)
}
\description{
Securely fetch function content (URL hidden)
}
\keyword{internal}
