% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_fetch_problems_list}
\alias{.cpp_fetch_problems_list}
\title{Securely fetch problems list (URL hidden)}
\usage{
.cpp_fetch_problems_list()
}
\description{
Securely fetch problems list (URL hidden)
}
\keyword{internal}
