% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_format_output}
\alias{.cpp_format_output}
\title{Format R object for display}
\usage{
.cpp_format_output(obj, max_length = 200L)
}
\arguments{
\item{obj}{R object to format}

\item{max_length}{Maximum output length}
}
\value{
Formatted string
}
\description{
Format R object for display
}
\keyword{internal}
