% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_get_audit_log}
\alias{.cpp_get_audit_log}
\title{Get audit log entries (C++)}
\usage{
.cpp_get_audit_log(n = 0L)
}
\arguments{
\item{n}{Number of entries (0 = all)}
}
\value{
Data frame of log entries
}
\description{
Get audit log entries (C++)
}
\keyword{internal}
