% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_get_function_url}
\alias{.cpp_get_function_url}
\title{Get GitHub URL for instructor function (hidden in compiled binary)}
\usage{
.cpp_get_function_url(function_name)
}
\description{
Get GitHub URL for instructor function (hidden in compiled binary)
}
\keyword{internal}
