% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_get_testdata_url}
\alias{.cpp_get_testdata_url}
\title{Get GitHub URL for test data (hidden in compiled binary)}
\usage{
.cpp_get_testdata_url()
}
\description{
Get GitHub URL for test data (hidden in compiled binary)
}
\keyword{internal}
