% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_get_type}
\alias{.cpp_get_type}
\title{Get R object type name}
\usage{
.cpp_get_type(obj)
}
\arguments{
\item{obj}{R object}
}
\value{
Type name string
}
\description{
Get R object type name
}
\keyword{internal}
