% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_has_internet}
\alias{.cpp_has_internet}
\title{Check internet connectivity}
\usage{
.cpp_has_internet()
}
\value{
Logical indicating if internet is available
}
\description{
Check internet connectivity
}
\keyword{internal}
