% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_log_security_event}
\alias{.cpp_log_security_event}
\title{Log security event (C++)}
\usage{
.cpp_log_security_event(event, details = "")
}
\arguments{
\item{event}{Event type}

\item{details}{Event details}
}
\description{
Log security event (C++)
}
\keyword{internal}
