% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_rate_limit_status}
\alias{.cpp_rate_limit_status}
\title{Get rate limit status (C++)}
\usage{
.cpp_rate_limit_status()
}
\value{
List with rate limit status
}
\description{
Get rate limit status (C++)
}
\keyword{internal}
