% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_secure_random_bytes}
\alias{.cpp_secure_random_bytes}
\title{Generate secure random bytes (C++)}
\usage{
.cpp_secure_random_bytes(n)
}
\arguments{
\item{n}{Number of bytes to generate}
}
\value{
Raw vector of random bytes
}
\description{
Generate secure random bytes (C++)
}
\keyword{internal}
