% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_secure_random_hex}
\alias{.cpp_secure_random_hex}
\title{Generate secure random hex string (C++)}
\usage{
.cpp_secure_random_hex(n)
}
\arguments{
\item{n}{Number of bytes (result will be 2*n characters)}
}
\value{
Hex-encoded string
}
\description{
Generate secure random hex string (C++)
}
\keyword{internal}
