% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_validate_function_name}
\alias{.cpp_validate_function_name}
\title{Validate function name}
\usage{
.cpp_validate_function_name(name)
}
\arguments{
\item{name}{Function name to validate}
}
\value{
Logical indicating if name is valid
}
\description{
Validate function name
}
\keyword{internal}
