% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail.R
\name{.fail}
\alias{.fail}
\title{Get a fail message}
\usage{
.fail(message = "Wrong. Please try again")
}
\arguments{
\item{message}{default message for wrong answers}
}
\value{
a message
}
\description{
Get a fail message
}
