% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{.guess_hw_name_from_path}
\alias{.guess_hw_name_from_path}
\title{Guess the homework name from a file path}
\usage{
.guess_hw_name_from_path(path, pattern = "hw\\\\d{2}")
}
\arguments{
\item{path}{character of length 1; file path}

\item{pattern}{character of length 1; regular expression used to extract the
homework name}
}
\value{
Returns the extracted homework name. Returns `NA` if no match was
  found
}
\description{
Guess the homework name from a file path
}
