% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_funs.R
\name{.insert_results_into_html}
\alias{.insert_results_into_html}
\title{Inserts homework check results into a rendered HTML file}
\usage{
.insert_results_into_html(result, path)
}
\arguments{
\item{result}{output from `.check_answers`}

\item{path}{character of length 1; path to the HTML file to insert the
results summary into.}
}
\description{
This function is used to insert the results of checking the
  homework into the top of the rendered HTML file. This provides a summary
  of what answers were correct.
}
