% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{.load_list}
\alias{.load_list}
\title{Load a .RData file as a list}
\usage{
.load_list(path)
}
\arguments{
\item{path}{character of length 1; path to the .RData file}
}
\value{
list containing the R objects
}
\description{
Given a .RData file, loads the objects it contains into a list.
}
