% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{.na}
\alias{.na}
\title{Get a invalid message}
\usage{
.na(message = "Invalid answer. Please check your work.")
}
\arguments{
\item{message}{default message for invalid answers}
}
\value{
a message
}
\description{
Get a invalid message
}
