% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name (passed by R)}

\item{pkgname}{Package name (passed by R)}
}
\description{
Displays welcome message when package is loaded via library() or require().
}
\details{
Message includes:
\itemize{
\item Package version
\item Quick start instructions (list_problems, preview_tests, autograder)
\item New features in this version
}

Uses packageStartupMessage() which can be suppressed with:
suppressPackageStartupMessages(library(autograder))
}
\keyword{internal}
