% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name (passed by R)}

\item{pkgname}{Package name (passed by R)}
}
\description{
Sets default package options when the package is loaded.
Initializes memoise cache and registers built-in comparison functions.
}
\keyword{internal}
