% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{.pass}
\alias{.pass}
\title{Generate a correct message}
\usage{
.pass(message = paste0("Correct! ", praise::praise()))
}
\arguments{
\item{message}{write a message different from default}
}
\value{
generate a message
}
\description{
Generate a correct message
}
