% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_funs.R
\name{.print_source_and_check}
\alias{.print_source_and_check}
\title{Prints the returned value of `.check_answers`}
\usage{
.print_source_and_check(result, width = 30)
}
\arguments{
\item{result}{output from `.check_answers`}
}
\value{
Returns a character vector of length one containing the printed
  results.
}
\description{
Prints the returned value of `.check_answers`
}
