% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{.process_key}
\alias{.process_key}
\title{Process key input}
\usage{
.process_key(key, path, pattern = "hw\\\\d{2}")
}
\arguments{
\item{key}{named list or character of length 1. If a named list, this value
is returned. If a character of length 1, this is used to try and retrieve
the key from the GitHub repository. If `NULL` it tries to guess the
homework name from `path`}

\item{path}{character of length 1; file path used to guess the homework name}

\item{pattern}{character of length 1; regular expression used to extract the
homework name from `path`.}
}
\description{
Processes the input to `key` in the way described in the
  `details` section of the `render_all` documentation. Throws an error if it
  is unable to guess the homework name.
}
