% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{.register_builtin_comparisons}
\alias{.register_builtin_comparisons}
\title{Register built-in comparison functions}
\usage{
.register_builtin_comparisons()
}
\description{
Register built-in comparison functions
}
\keyword{internal}
