% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answers.R
\name{.run_render}
\alias{.run_render}
\title{Renders an R or Rmarkdown file}
\usage{
.run_render(path, env, out_dir)
}
\arguments{
\item{path}{character of length 1; path to the file to render}

\item{env}{environment; the objects created by the script will be saved in
this environment}

\item{out_dir}{character of length(1); path to the directory in which to save
the rendered file.}
}
\description{
Renders an R or Rmarkdown file while store the created objects
  in a user-provided environment.
}
