% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_grid.R
\name{.score_grid}
\alias{.score_grid}
\title{Create a grid with the score of each question}
\usage{
.score_grid()
}
\value{
a grid number and color-coded scores
}
\description{
Create a grid with the score of each question
}
\details{
\if{html}{\out{<div class="sourceCode {asciicast cli-cli}">}}\preformatted{.scores <- rep(-1, times = 3)   # Put total number of questions as `times` argument
.score_grid()
}\if{html}{\out{</div>}}
}
