% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_print.R
\name{.score_print}
\alias{.score_print}
\title{Print complete scores!}
\usage{
.score_print()
}
\value{
final scores
}
\description{
Print complete scores!
}
\details{
\if{html}{\out{<div class="sourceCode {asciicast cli-print}">}}\preformatted{.scores <- rep(-1, times = 3)   # Put total number of questions as `times` argument
.score_print()
}\if{html}{\out{</div>}}
}
