% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_funs.R
\name{.shorten_output}
\alias{.shorten_output}
\title{Shortens the printed version of an object}
\usage{
.shorten_output(x, n)
}
\arguments{
\item{x}{object to print}

\item{n}{integer; maximum number of characters}
}
\value{
character vector of length 1 with the shortened output.
}
\description{
Given an object and a number of characters, captures the output
  from printing the object and then shortens it to the desired number of
  characters (if necessary).
}
