% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answers.R
\name{.source_and_check_items}
\alias{.source_and_check_items}
\title{Runs a homework script and compares the answers to the answer key}
\usage{
.source_and_check_items(
  path,
  key_items,
  key_objects,
  source_fun = .run_source,
  ...
)
}
\arguments{
\item{path}{character of length 1; path to the script to run}

\item{source_fun}{function for running the script. Must accept at least two
parameters - the first should be a character of length 1 that contains the
path to the script to run. The second should be an environment, and all
objects created in the script should be saved in this environment. If the
source script doesn't run, this function should return a `try-error`.
`.run_source` and `.run_render` have been designed to work with this
function.}

\item{...}{additional arguments passed to `source_fun`}

\item{key}{named list containing the objects that represent the correct
answers.}
}
\value{
list with three elements:
* `result`: output from `.compare_objects` (data frame)
* `source_error`: character; error message from running the script. `NA` if
  no error occurred.
* `answers`: list containing the objects created from running the homework
  script. Only contains the objects whose name appears in `key`.
* `key`: the passed to the `key` parameter
}
\description{
Runs a homework script and compares the answers to the answer key
}
