% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_script}
\alias{example_script}
\title{Get example script.}
\usage{
example_script(name)
}
\arguments{
\item{name}{character string that is the name of the script}
}
\value{
a character string which contains a full file location and name
}
\description{
Get example script.
}
\examples{
example_script("generic_script")
}
