% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{extract_instructor_function}
\alias{extract_instructor_function}
\title{Extract instructor function from environment}
\usage{
extract_instructor_function(instructor_env, function_name)
}
\description{
Extract instructor function from environment
}
\keyword{internal}
