% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{extract_test_cases}
\alias{extract_test_cases}
\title{Extract and validate test cases}
\usage{
extract_test_cases(instructor_env, function_name)
}
\description{
Extract and validate test cases
}
\keyword{internal}
