% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parsed_tibble.R
\name{file_validation}
\alias{file_validation}
\title{Validate a file.}
\usage{
file_validation(filename)
}
\arguments{
\item{filename}{a character string which is a file name}
}
\description{
Checks whether file exists (error) and
check whether file is an R script (warning)
}
\examples{
example_script("generic_script") |>
  file_validation()
}
