% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{format_output}
\alias{format_output}
\title{Improved format output with smart truncation}
\usage{
format_output(obj, max_length = 200, preserve_structure = TRUE)
}
\description{
Improved format output with smart truncation
}
\keyword{internal}
