% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_student.R
\name{generate_student}
\alias{generate_student}
\title{Generate student's and full version of the homework}
\usage{
generate_student()
}
\description{
This function takes a pre-defined HW template and generates a full
homework with an autograding system
}
