% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{get_key}
\alias{get_key}
\title{Read a homework key from the GitHub repository.}
\usage{
get_key(name)
}
\arguments{
\item{name}{name of the homework key. Can include the .RData file extension
but doesn't need to.}
}
\value{
Returns the key as a named list.
}
\description{
The homework keys can be found in this repository:
https://github.com/drkfrnd/STAT212.
}
