% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parsed_tibble.R
\name{get_parsed_tibble}
\alias{get_parsed_tibble}
\title{Parse a script into a \code{tibble}.}
\usage{
get_parsed_tibble(filename, file_validation = TRUE)
}
\arguments{
\item{filename}{a character string which is a file name}

\item{file_validation}{if \code{TRUE} calls \code{file_validation} function (default). If \code{FALSE}, skips file validation; defaults to system validation.}
}
\value{
a tibble of parsed code on which to operate
}
\description{
Parses file and puts it in a tibble, with file validation
}
\examples{
example_script("generic_script") |>
  get_parsed_tibble()
}
