% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/validation.R
\name{get_test_count}
\alias{get_test_count}
\title{Get test count from test data}
\usage{
get_test_count(test_data)
}
\arguments{
\item{test_data}{List containing test case configuration}
}
\value{
Number of tests, or -1 if invalid

Number of tests, or -1L if invalid
}
\description{
Get test count from test data

Get test count from test data (pure R)
}
\keyword{internal}
