% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeR.R
\name{gradeR}
\alias{gradeR}
\title{Autograder}
\usage{
gradeR(student_answers, reference_answers, maximum_grade = 10)
}
\arguments{
\item{student_answers}{Answers provided by the student in the form of a list that matches the order as the one in the \code{reference_answers}}

\item{reference_answers}{Reference answers constructed using the files under /tests}

\item{maximum_grade}{Maximum grade}
}
\description{
This function combines student's answers, reference answers, and a maximum number
of points to grade a particular set of questions.
}
