% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework_template.R
\name{homework_template}
\alias{homework_template}
\title{Create a homework template for gradeR}
\usage{
homework_template(path)
}
\arguments{
\item{path}{Where should the sample directory be exported to?}
}
\description{
This function generates the structure of a basic autograding system.
}
