% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{key_fun}
\alias{key_fun}
\title{Get a function for appending to a key}
\usage{
key_fun()
}
\value{
Returns a function with the following function signature:
`function(name, ...)`. The parameters are:
* `name`: character of length 1; name to assign to this item. If this parameter
  is missing no item is added and the key is returned.
* `...`: arguments passed to `item()`.
}
\description{
This function is intended to be used when creating a key. This
function returns a function, and you can then use that function to add to the
key. The goal is to minimize typing and make it quick and easy to add items
to a key.
}
\examples{
k <- key_fun()
x <- 1
k("q1", x, pts = 1)
y <- "hello"
k("q2", y, pts = 1.5)
k() # return the key
# saveRDS(k(), "mykey.rds")
}
