% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{length_mismatch_error}
\alias{length_mismatch_error}
\title{Custom error for length mismatch}
\usage{
length_mismatch_error(expected_length, actual_length, call = NULL)
}
\arguments{
\item{expected_length}{The length that was expected}

\item{actual_length}{The length that was returned}

\item{call}{The call that generated the error}
}
\value{
An S3 object of class c("autograder_length_mismatch_error", "autograder_error", "error", "condition")
}
\description{
Creates a structured error when the student's output length doesn't match
the expected length.
}
\keyword{internal}
