% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{list_problems}
\alias{list_problems}
\title{List available problems}
\usage{
list_problems()
}
\description{
Shows all available functions that can be tested with autograder.
}
\examples{
\dontrun{
list_problems()
}

}
