% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_funs.R
\name{load_key}
\alias{load_key}
\title{Load a .RData file as a list}
\usage{
load_key(path)
}
\arguments{
\item{path}{character of length 1; path to the .RData file}
}
\value{
list containing the R objects
}
\description{
Given an answer key stored as a .RData file, loads the objects
  it contains into a list.
}
