% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{pipe_counter}
\alias{pipe_counter}
\title{Count \code{base} and \code{magrittr} pipe usages}
\usage{
pipe_counter(filename)
}
\arguments{
\item{filename}{a character string which is a file name}
}
\value{
a tibble with counts of each pipe type, \verb{|>} and \verb{\%>\%}
}
\description{
This function takes an R script and
outputs a tibble with pipe counts by type
always outputs a count for both
}
\examples{
example_script("generic_tidyverse_script") |>
  pipe_counter()
}
