% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{print.openmp_status}
\alias{print.openmp_status}
\title{Print OpenMP Status}
\usage{
\method{print}{openmp_status}(x, ...)
}
\arguments{
\item{x}{An openmp_status object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print OpenMP Status
}
\keyword{internal}
