% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{provide_feedback}
\alias{provide_feedback}
\title{Analyze error and provide helpful hints}
\usage{
provide_feedback(student_out, expected_out, input_args, hint = NULL)
}
\description{
Analyze error and provide helpful hints
}
\keyword{internal}
