% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_funs.R
\name{render_all}
\alias{render_all}
\title{Renders multiple homework assignments}
\usage{
render_all(
  in_dir,
  out_dir,
  key = NULL,
  summary_csv = TRUE,
  pattern = "[.](R|Rmd)$"
)
}
\arguments{
\item{in_dir}{character of length 1; path to directory containing R or Rmd
scripts.}

\item{out_dir}{character of length 1; path to directory in which to output
the HTML files}

\item{key}{One of the following: `NULL`, a character of length 1, or a named
list. See details for more.}

\item{summary_csv}{logical of length 1; if `TRUE`, outputs a CSV with a
summary of the results for each script.}

\item{pattern}{character of length 1; regular expression used to select
the scripts in `out_dir`}
}
\description{
Runs all R or Rmd files in a directory and creates an HTML file
  for each script that contains the rendered code as well as a summary of
  which answers were correct/incorrect. The function also optionally outputs
  a CSV giving a summary of the results.
}
\details{
There are three possible kinds of inputs to `key`:
  * character of length 1: in this case `key` is taken to be the name of a
key stored in the GitHub repository (https://github.com/drkfrnd/STAT212). The
key is read from GitHub and used to check the answers
  * `NULL`: if `key` is `NULL` (the default) the name of the key is guessed
from `in_dir` by trying to extract a string matching the regular expression
"hw\\d{2}". If found, uses this as the key name. Otherwise an error is
thrown.
  * named list containing the answers (as in `render_hw()`).
}
