% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_funs.R
\name{render_hw}
\alias{render_hw}
\title{Render homework to an HTML file}
\usage{
render_hw(path, key, out_dir = NULL)
}
\arguments{
\item{path}{character of length 1; path to the homework script}

\item{key}{named list containing the objects that represent the correct
answers.}

\item{out_dir}{character of the length 1; path to the directory in which to
save the resulting HTML. If `NULL` outputs to the current working
directory.}
}
\description{
Renders a homework script to an HTML file. The resulting HTML
  file includes output at the top that summarizes the result of comparing the
  homework answers to the key
}
