% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{return_checker}
\alias{return_checker}
\title{Checks whether functions call \code{return}.}
\usage{
return_checker(filename)
}
\arguments{
\item{filename}{a character string which is a file name}
}
\value{
a vector of functions that are missing a \code{return} call
}
\description{
This function takes an R script and
outputs a vector of function names that don't have return calls
}
\examples{
example_script("generic_script") |>
  return_checker()
}
