% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{secure_random_hex}
\alias{secure_random_hex}
\title{Generate cryptographically secure random hex string}
\usage{
secure_random_hex(n)
}
\arguments{
\item{n}{Number of bytes to generate (result will be 2*n characters)}
}
\value{
Hex-encoded string
}
\description{
Generates a hex-encoded string of secure random bytes.
}
\keyword{internal}
