% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{show_character_diff}
\alias{show_character_diff}
\title{Show character vector differences}
\usage{
show_character_diff(expected, actual, max_show)
}
\description{
Show character vector differences
}
\keyword{internal}
