% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{show_dataframe_diff}
\alias{show_dataframe_diff}
\title{Show data frame differences}
\usage{
show_dataframe_diff(expected, actual, max_show, tolerance)
}
\description{
Show data frame differences
}
\keyword{internal}
