% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{show_numeric_diff}
\alias{show_numeric_diff}
\title{Show numeric vector differences}
\usage{
show_numeric_diff(expected, actual, max_show, tolerance)
}
\description{
Show numeric vector differences
}
\keyword{internal}
