% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{test_execution_error}
\alias{test_execution_error}
\title{Custom error for test execution}
\usage{
test_execution_error(message, test_number, call = NULL)
}
\description{
Custom error for test execution
}
\keyword{internal}
