% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{autograder}
\alias{autograder}
\title{Run autograder by function name}
\usage{
autograder(function_name)
}
\arguments{
\item{function_name}{Character. Name of the function to test.}
}
\description{
Run autograder by function name
}
