% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answers.R
\name{check_hw}
\alias{check_hw}
\title{Check homework answers}
\usage{
check_hw(path, key = NULL, verbose = TRUE)
}
\arguments{
\item{path}{character of length 1; path to an R script to check.}

\item{key}{One of the following: `NULL`, a character of length 1, or a named
list. See details for more.}

\item{verbose}{logical of length 1; if `TRUE`, prints the result of checking
the homework assignment to the console.}
}
\value{
invisibly returns the returned value from `.check_answers`
}
\description{
Runs an R script and compares the objects created in the script
  to an answer key. Note that the R script is run in a new session, so if the
  script references objects not created in the script, an error will be
  thrown.
}
