% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_compare_fast}
\alias{.cpp_compare_fast}
\title{Fast comparison of R objects using C++}
\usage{
.cpp_compare_fast(obj1, obj2, tolerance = 1e-10)
}
\description{
Fast comparison of R objects using C++
}
\keyword{internal}
