% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cpp_compare_identical}
\alias{.cpp_compare_identical}
\title{Legacy comparison function (kept for compatibility)}
\usage{
.cpp_compare_identical(obj1, obj2)
}
\description{
Legacy comparison function (kept for compatibility)
}
\keyword{internal}
