% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_fetch_function_content}
\alias{.cpp_fetch_function_content}
\title{Fetch function content from repository}
\usage{
.cpp_fetch_function_content(function_name)
}
\arguments{
\item{function_name}{Name of function to fetch}
}
\value{
Character vector of file lines
}
\description{
Fetch function content from repository
}
\keyword{internal}
