% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.cpp_fetch_problems_list}
\alias{.cpp_fetch_problems_list}
\title{Fetch problems list from repository}
\usage{
.cpp_fetch_problems_list()
}
\value{
Character vector of problems file content
}
\description{
Fetch problems list from repository
}
\keyword{internal}
