% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answers.R
\name{.run_source}
\alias{.run_source}
\title{Run a script and store the created objects}
\usage{
.run_source(path, env, ...)
}
\arguments{
\item{path}{character of length 1; path to the R script (or .Rmd)}

\item{env}{environment; the objects created by the script will be saved in
this environment}

\item{...}{ignored}
}
\value{
If `source()` runs successfully, returns the output of `source()`.
  Otherwise returns a `try-error`.
}
\description{
Sources an R script and saves the objects created by the script
  into a user-provided environment.
}
