% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{fetch_instructor_code}
\alias{fetch_instructor_code}
\title{Securely fetch and load instructor code}
\usage{
fetch_instructor_code(function_name)
}
\description{
Securely fetch and load instructor code
}
\keyword{internal}
