% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{format_output}
\alias{format_output}
\title{Format output for display (truncate if too long)}
\usage{
format_output(obj, max_length = 100)
}
\description{
Format output for display (truncate if too long)
}
\keyword{internal}
