% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{function_not_found_error}
\alias{function_not_found_error}
\title{Custom error for missing functions}
\usage{
function_not_found_error(function_name, call = NULL)
}
\description{
Custom error for missing functions
}
\keyword{internal}
