% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_funs.R
\name{render_no_code}
\alias{render_no_code}
\title{Render a script without the code}
\usage{
render_no_code(path, output_format = "all", ...)
}
\arguments{
\item{path}{character of length 1; path to the R or Rmd file}

\item{output_file}{character of length 1; path to the output file}
}
\description{
Renders a script while ignoring the code
}
