% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{return_checker}
\alias{return_checker}
\title{return_checker
This function takes an R script and
outputs a vector of function names that don't have return calls}
\usage{
return_checker(filename)
}
\arguments{
\item{filename}{a character string which is a file name}
}
\value{
a vector of functions that are missing a \code{return} call
}
\description{
return_checker
This function takes an R script and
outputs a vector of function names that don't have return calls
}
\examples{
try(return_checker())
}
