% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{run_tests_parallel}
\alias{run_tests_parallel}
\title{Run tests in parallel}
\usage{
run_tests_parallel(
  student_fun,
  instructor_fun,
  test_data,
  tolerance,
  use_parallel = TRUE
)
}
\description{
Run tests in parallel
}
\keyword{internal}
