% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{run_tests_sequential}
\alias{run_tests_sequential}
\title{Run tests sequentially}
\usage{
run_tests_sequential(student_fun, instructor_fun, test_data, tolerance)
}
\description{
Run tests sequentially
}
\keyword{internal}
