% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{validate_test_cases}
\alias{validate_test_cases}
\title{Validate test case structure}
\usage{
validate_test_cases(test_data, function_name)
}
\description{
Validate test case structure
}
\keyword{internal}
